% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_directories.R
\docType{class}
\name{AccountDirectories}
\alias{AccountDirectories}
\title{AccountDirectories}
\format{An \code{R6Class} generator object}
\usage{
AccountDirectories
}
\description{
AccountDirectories Class
}
\section{Fields}{

\describe{
\item{\code{accounts}}{list(character)   [optional]}

\item{\code{directories}}{list(character)   [optional]}
}}

\keyword{datasets}
