% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\docType{class}
\name{ApiClient}
\alias{ApiClient}
\title{ApiClient}
\format{An \code{R6Class} generator object}
\usage{
ApiClient
}
\description{
ApiClient Class
}
\details{
ApiClient Class

Generic API client for OpenAPI client library builds.
OpenAPI generic API client. This client handles the client-
server communication, and is invariant across implementations. Specifics of
the methods and models for each application are generated from the OpenAPI Generator
templates.

NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
Ref: https://openapi-generator.tech
Do not edit the class manually.
}
\section{Fields}{

\describe{
\item{\code{basePath}}{Base url}

\item{\code{userAgent}}{Default user agent}

\item{\code{username}}{Username for HTTP basic authentication}

\item{\code{password}}{Password for HTTP basic authentication}

\item{\code{timeout}}{Default timeout in seconds}
}}

\keyword{datasets}
