% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_exception.R
\docType{class}
\name{ApiException}
\alias{ApiException}
\title{ApiException}
\format{An \code{R6Class} generator object}
\usage{
ApiException
}
\description{
ApiException Class
}
\section{Fields}{

\describe{
\item{\code{status}}{Status of the ApiException}

\item{\code{reason}}{Reason of the ApiException}

\item{\code{body}}{Body of the http response}

\item{\code{headers}}{Headers of the http response}
}}

\keyword{datasets}
