% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\docType{class}
\name{Calculation}
\alias{Calculation}
\title{Calculation}
\format{An \code{R6Class} generator object}
\usage{
Calculation
}
\description{
Calculation Class
}
\section{Fields}{

\describe{
\item{\code{pa}}{\link[factset.analyticsapi.engines.v2:PACalculationParameters]{   named list(PACalculationParameters)   }  [optional]}

\item{\code{spar}}{\link[factset.analyticsapi.engines.v2:SPARCalculationParameters]{   named list(SPARCalculationParameters)   }  [optional]}

\item{\code{vault}}{\link[factset.analyticsapi.engines.v2:VaultCalculationParameters]{   named list(VaultCalculationParameters)   }  [optional]}
}}

\keyword{datasets}
