% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_status.R
\docType{class}
\name{CalculationStatus}
\alias{CalculationStatus}
\title{CalculationStatus}
\format{An \code{R6Class} generator object}
\usage{
CalculationStatus
}
\description{
CalculationStatus Class
}
\section{Fields}{

\describe{
\item{\code{status}}{character  [optional]}

\item{\code{points}}{integer  [optional]}

\item{\code{pa}}{\link[factset.analyticsapi.engines.v2:CalculationUnitStatus]{   named list(CalculationUnitStatus)   }  [optional]}

\item{\code{spar}}{\link[factset.analyticsapi.engines.v2:CalculationUnitStatus]{   named list(CalculationUnitStatus)   }  [optional]}

\item{\code{vault}}{\link[factset.analyticsapi.engines.v2:CalculationUnitStatus]{   named list(CalculationUnitStatus)   }  [optional]}
}}

\keyword{datasets}
