% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_status_summary.R
\docType{class}
\name{CalculationStatusSummary}
\alias{CalculationStatusSummary}
\title{CalculationStatusSummary}
\format{An \code{R6Class} generator object}
\usage{
CalculationStatusSummary
}
\description{
CalculationStatusSummary Class
}
\section{Fields}{

\describe{
\item{\code{status}}{character  [optional]}

\item{\code{points}}{integer  [optional]}

\item{\code{requestTime}}{character  [optional]}

\item{\code{lastPollTime}}{character  [optional]}
}}

\keyword{datasets}
