% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_unit_status.R
\docType{class}
\name{CalculationUnitStatus}
\alias{CalculationUnitStatus}
\title{CalculationUnitStatus}
\format{An \code{R6Class} generator object}
\usage{
CalculationUnitStatus
}
\description{
CalculationUnitStatus Class
}
\section{Fields}{

\describe{
\item{\code{status}}{character  [optional]}

\item{\code{points}}{integer  [optional]}

\item{\code{error}}{character  [optional]}

\item{\code{result}}{character  [optional]}
}}

\keyword{datasets}
