% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\docType{class}
\name{Column}
\alias{Column}
\title{Column}
\format{An \code{R6Class} generator object}
\usage{
Column
}
\description{
Column Class
}
\section{Fields}{

\describe{
\item{\code{defaultstatisticsids}}{list(character)   [optional]}

\item{\code{name}}{character  [optional]}

\item{\code{directory}}{character  [optional]}

\item{\code{category}}{character  [optional]}
}}

\keyword{datasets}
