% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_statistic.R
\docType{class}
\name{ColumnStatistic}
\alias{ColumnStatistic}
\title{ColumnStatistic}
\format{An \code{R6Class} generator object}
\usage{
ColumnStatistic
}
\description{
ColumnStatistic Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character  [optional]}
}}

\keyword{datasets}
