% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_summary.R
\docType{class}
\name{ColumnSummary}
\alias{ColumnSummary}
\title{ColumnSummary}
\format{An \code{R6Class} generator object}
\usage{
ColumnSummary
}
\description{
ColumnSummary Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character  [optional]}

\item{\code{directory}}{character  [optional]}

\item{\code{category}}{character  [optional]}
}}

\keyword{datasets}
