% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_account.R
\docType{class}
\name{ComponentAccount}
\alias{ComponentAccount}
\title{ComponentAccount}
\format{An \code{R6Class} generator object}
\usage{
ComponentAccount
}
\description{
ComponentAccount Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character  [optional]}

\item{\code{name}}{character  [optional]}

\item{\code{holdingsmode}}{character  [optional]}
}}

\keyword{datasets}
