% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_benchmark.R
\docType{class}
\name{ComponentBenchmark}
\alias{ComponentBenchmark}
\title{ComponentBenchmark}
\format{An \code{R6Class} generator object}
\usage{
ComponentBenchmark
}
\description{
ComponentBenchmark Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character  [optional]}

\item{\code{name}}{character  [optional]}

\item{\code{holdingsmode}}{character  [optional]}
}}

\keyword{datasets}
