% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_account.R
\docType{class}
\name{ConfigurationAccount}
\alias{ConfigurationAccount}
\title{ConfigurationAccount}
\format{An \code{R6Class} generator object}
\usage{
ConfigurationAccount
}
\description{
ConfigurationAccount Class
}
\section{Fields}{

\describe{
\item{\code{benchmarkCode}}{character  [optional]}

\item{\code{benchmarkName}}{character  [optional]}

\item{\code{maxEndDate}}{character  [optional]}

\item{\code{minStartDate}}{character  [optional]}

\item{\code{lockingDate}}{character  [optional]}

\item{\code{name}}{character  [optional]}
}}

\keyword{datasets}
