% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\docType{class}
\name{Currency}
\alias{Currency}
\title{Currency}
\format{An \code{R6Class} generator object}
\usage{
Currency
}
\description{
Currency Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character  [optional]}
}}

\keyword{datasets}
