% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_parameters_summary.R
\docType{class}
\name{DateParametersSummary}
\alias{DateParametersSummary}
\title{DateParametersSummary}
\format{An \code{R6Class} generator object}
\usage{
DateParametersSummary
}
\description{
DateParametersSummary Class
}
\section{Fields}{

\describe{
\item{\code{startdate}}{character  [optional]}

\item{\code{enddate}}{character  [optional]}
}}

\keyword{datasets}
