% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_directories.R
\docType{class}
\name{DocumentDirectories}
\alias{DocumentDirectories}
\title{DocumentDirectories}
\format{An \code{R6Class} generator object}
\usage{
DocumentDirectories
}
\description{
DocumentDirectories Class
}
\section{Fields}{

\describe{
\item{\code{documents}}{list(character)   [optional]}

\item{\code{directories}}{list(character)   [optional]}
}}

\keyword{datasets}
