% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\docType{class}
\name{Frequency}
\alias{Frequency}
\title{Frequency}
\format{An \code{R6Class} generator object}
\usage{
Frequency
}
\description{
Frequency Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character  [optional]}
}}

\keyword{datasets}
