% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\docType{class}
\name{Group}
\alias{Group}
\title{Group}
\format{An \code{R6Class} generator object}
\usage{
Group
}
\description{
Group Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character  [optional]}

\item{\code{directory}}{character  [optional]}

\item{\code{category}}{character  [optional]}
}}

\keyword{datasets}
