% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_calculation_column.R
\docType{class}
\name{PACalculationColumn}
\alias{PACalculationColumn}
\title{PACalculationColumn}
\format{An \code{R6Class} generator object}
\usage{
PACalculationColumn
}
\description{
PACalculationColumn Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character  [optional]}

\item{\code{statistics}}{list(character)   [optional]}
}}

\keyword{datasets}
