% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_calculation_group.R
\docType{class}
\name{PACalculationGroup}
\alias{PACalculationGroup}
\title{PACalculationGroup}
\format{An \code{R6Class} generator object}
\usage{
PACalculationGroup
}
\description{
PACalculationGroup Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character  [optional]}
}}

\keyword{datasets}
