% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_calculation_parameters.R
\docType{class}
\name{PACalculationParameters}
\alias{PACalculationParameters}
\title{PACalculationParameters}
\format{An \code{R6Class} generator object}
\usage{
PACalculationParameters
}
\description{
PACalculationParameters Class
}
\section{Fields}{

\describe{
\item{\code{componentid}}{character}

\item{\code{accounts}}{\link[factset.analyticsapi.engines.v2:PAIdentifier]{   list(PAIdentifier)   }  [optional]}

\item{\code{benchmarks}}{\link[factset.analyticsapi.engines.v2:PAIdentifier]{   list(PAIdentifier)   }  [optional]}

\item{\code{dates}}{\link[factset.analyticsapi.engines.v2:PADateParameters]{  PADateParameters  }  [optional]}

\item{\code{groups}}{\link[factset.analyticsapi.engines.v2:PACalculationGroup]{   list(PACalculationGroup)   }  [optional]}

\item{\code{currencyisocode}}{character  [optional]}

\item{\code{columns}}{\link[factset.analyticsapi.engines.v2:PACalculationColumn]{   list(PACalculationColumn)   }  [optional]}
}}

\keyword{datasets}
