% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_component.R
\docType{class}
\name{PAComponent}
\alias{PAComponent}
\title{PAComponent}
\format{An \code{R6Class} generator object}
\usage{
PAComponent
}
\description{
PAComponent Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character  [optional]}

\item{\code{accounts}}{\link[factset.analyticsapi.engines.v2:PAIdentifier]{   list(PAIdentifier)   }  [optional]}

\item{\code{benchmarks}}{\link[factset.analyticsapi.engines.v2:PAIdentifier]{   list(PAIdentifier)   }  [optional]}

\item{\code{currencyisocode}}{character  [optional]}

\item{\code{dates}}{\link[factset.analyticsapi.engines.v2:PADateParameters]{  PADateParameters  }  [optional]}

\item{\code{snapshot}}{character  [optional]}

\item{\code{name}}{character  [optional]}

\item{\code{category}}{character  [optional]}
}}

\keyword{datasets}
