% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_date_parameters.R
\docType{class}
\name{PADateParameters}
\alias{PADateParameters}
\title{PADateParameters}
\format{An \code{R6Class} generator object}
\usage{
PADateParameters
}
\description{
PADateParameters Class
}
\section{Fields}{

\describe{
\item{\code{startdate}}{character  [optional]}

\item{\code{enddate}}{character}

\item{\code{frequency}}{character}
}}

\keyword{datasets}
