% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_identifier.R
\docType{class}
\name{PAIdentifier}
\alias{PAIdentifier}
\title{PAIdentifier}
\format{An \code{R6Class} generator object}
\usage{
PAIdentifier
}
\description{
PAIdentifier Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character}

\item{\code{holdingsmode}}{character  [optional]}
}}

\keyword{datasets}
