% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spar_calculation_parameters.R
\docType{class}
\name{SPARCalculationParameters}
\alias{SPARCalculationParameters}
\title{SPARCalculationParameters}
\format{An \code{R6Class} generator object}
\usage{
SPARCalculationParameters
}
\description{
SPARCalculationParameters Class
}
\section{Fields}{

\describe{
\item{\code{componentid}}{character}

\item{\code{accounts}}{\link[factset.analyticsapi.engines.v2:SPARIdentifier]{   list(SPARIdentifier)   }  [optional]}

\item{\code{benchmark}}{\link[factset.analyticsapi.engines.v2:SPARIdentifier]{  SPARIdentifier  }  [optional]}

\item{\code{dates}}{\link[factset.analyticsapi.engines.v2:SPARDateParameters]{  SPARDateParameters  }  [optional]}
}}

\keyword{datasets}
