% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spar_date_parameters.R
\docType{class}
\name{SPARDateParameters}
\alias{SPARDateParameters}
\title{SPARDateParameters}
\format{An \code{R6Class} generator object}
\usage{
SPARDateParameters
}
\description{
SPARDateParameters Class
}
\section{Fields}{

\describe{
\item{\code{startdate}}{character}

\item{\code{enddate}}{character}

\item{\code{frequency}}{character}
}}

\keyword{datasets}
