% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spar_identifier.R
\docType{class}
\name{SPARIdentifier}
\alias{SPARIdentifier}
\title{SPARIdentifier}
\format{An \code{R6Class} generator object}
\usage{
SPARIdentifier
}
\description{
SPARIdentifier Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character}

\item{\code{returntype}}{character  [optional]}

\item{\code{prefix}}{character  [optional]}
}}

\keyword{datasets}
