% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_calculation_parameters.R
\docType{class}
\name{VaultCalculationParameters}
\alias{VaultCalculationParameters}
\title{VaultCalculationParameters}
\format{An \code{R6Class} generator object}
\usage{
VaultCalculationParameters
}
\description{
VaultCalculationParameters Class
}
\section{Fields}{

\describe{
\item{\code{componentid}}{character}

\item{\code{account}}{\link[factset.analyticsapi.engines.v2:VaultIdentifier]{  VaultIdentifier  }}

\item{\code{dates}}{\link[factset.analyticsapi.engines.v2:VaultDateParameters]{  VaultDateParameters  }  [optional]}

\item{\code{configid}}{character}
}}

\keyword{datasets}
