% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_component.R
\docType{class}
\name{VaultComponent}
\alias{VaultComponent}
\title{VaultComponent}
\format{An \code{R6Class} generator object}
\usage{
VaultComponent
}
\description{
VaultComponent Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character  [optional]}

\item{\code{account}}{\link[factset.analyticsapi.engines.v2:VaultIdentifier]{  VaultIdentifier  }  [optional]}

\item{\code{benchmark}}{\link[factset.analyticsapi.engines.v2:VaultIdentifier]{  VaultIdentifier  }  [optional]}

\item{\code{currencyisocode}}{character  [optional]}

\item{\code{dates}}{\link[factset.analyticsapi.engines.v2:VaultDateParameters]{  VaultDateParameters  }  [optional]}

\item{\code{snapshot}}{character  [optional]}

\item{\code{name}}{character  [optional]}

\item{\code{category}}{character  [optional]}
}}

\keyword{datasets}
