% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_configuration.R
\docType{class}
\name{VaultConfiguration}
\alias{VaultConfiguration}
\title{VaultConfiguration}
\format{An \code{R6Class} generator object}
\usage{
VaultConfiguration
}
\description{
VaultConfiguration Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character  [optional]}

\item{\code{accounts}}{\link[factset.analyticsapi.engines.v2:ConfigurationAccount]{   named list(ConfigurationAccount)   }  [optional]}
}}

\keyword{datasets}
