% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_configuration_summary.R
\docType{class}
\name{VaultConfigurationSummary}
\alias{VaultConfigurationSummary}
\title{VaultConfigurationSummary}
\format{An \code{R6Class} generator object}
\usage{
VaultConfigurationSummary
}
\description{
VaultConfigurationSummary Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character  [optional]}
}}

\keyword{datasets}
