% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_date_parameters.R
\docType{class}
\name{VaultDateParameters}
\alias{VaultDateParameters}
\title{VaultDateParameters}
\format{An \code{R6Class} generator object}
\usage{
VaultDateParameters
}
\description{
VaultDateParameters Class
}
\section{Fields}{

\describe{
\item{\code{startdate}}{character  [optional]}

\item{\code{enddate}}{character}

\item{\code{frequency}}{character}
}}

\keyword{datasets}
