% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_identifier.R
\docType{class}
\name{VaultIdentifier}
\alias{VaultIdentifier}
\title{VaultIdentifier}
\format{An \code{R6Class} generator object}
\usage{
VaultIdentifier
}
\description{
VaultIdentifier Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character}
}}

\keyword{datasets}
